source/requirements-dev.txt
avro==1.10.2
black
boto3
aws_cdk.core>=1.120.0
aws_cdk.aws_stepfunctions_tasks>=1.120.0
aws_solutions_constructs.aws_lambda_sns>=1.120.0
requests==2.24.0
crhelper==2.0.6
cronex==0.1.3.1
moto==2.0.8
parsedatetime==2.6
pytest
pytest-cov>=2.11.1
pytest-env>=0.6.2
pytest-mock>=3.5.1
tenacity>=8.0.1
-e cdk_solution_helper_py/helpers_cdk
-e cdk_solution_helper_py/helpers_common
aws-lambda-powertools==1.15.0
docker==5.0.0
-e infrastructure